refactor function locations to be more descriptive
