+++ src/asf_tools/hydrosar/hand/calculate.py
@@ -15,7 +15,7 @@ import rasterio.mask
 from pysheds.sgrid import sGrid
 from shapely.geometry import GeometryCollection, shape
 
-from asf_tools.composite import write_cog
+from asf_tools.raster import write_cog
 from asf_tools.dem import prepare_dem_vrt
 
 log = logging.getLogger(__name__)
