+++ src/asf_tools/raster.py
@@ -1,10 +1,13 @@
 import logging
 import warnings
 from pathlib import Path
-from typing import Literal, Union
+from tempfile import NamedTemporaryFile
+from typing import Literal, Union, List
 
 import numpy as np
 from osgeo import gdal
+
+from asf_tools.util import epsg_to_wkt
 
 gdal.UseExceptions()
 log = logging.getLogger(__name__)
@@ -57,4 +60,56 @@ def read_as_masked_array(raster: Union[str, Path], band: int = 1) -> np.ma.Maske
     nodata = band.GetNoDataValue()
     if nodata is not None:
         return np.ma.masked_values(data, nodata)
+    del ds  # How to close w/ gdal
     return data
+
+
+def read_as_array(raster: str, band: int = 1) -> np.array:
+    """Reads data from a raster image into memory
+
+    Args:
+        raster: The file path to a raster image
+        band: The raster band to read
+
+    Returns:
+        data: The raster pixel data as a numpy array
+    """
+    log.debug(f'Reading raster values from {raster}')
+    ds = gdal.Open(raster)
+    data = ds.GetRasterBand(band).ReadAsArray()
+    del ds  # How to close w/ gdal
+    return data
+
+
+def write_cog(file_name: Union[str, Path], data: np.ndarray, transform: List[float], epsg_code: int,
+              dtype=gdal.GDT_Float32, nodata_value=None):
+    """Creates a Cloud Optimized GeoTIFF
+
+    Args:
+        file_name: The output file name
+        data: The raster data
+        transform: The geotransform for the output GeoTIFF
+        epsg_code: The integer EPSG code for the output GeoTIFF projection
+        dtype: The pixel data type for the output GeoTIFF
+        nodata_value: The NODATA value for the output Geotiff
+
+    Returns:
+        file_name: The output file name
+    """
+    log.info(f'Creating {file_name}')
+
+    with NamedTemporaryFile() as temp_file:
+        driver = gdal.GetDriverByName('GTiff')
+        temp_geotiff = driver.Create(temp_file.name, data.shape[1], data.shape[0], 1, dtype)
+        temp_geotiff.GetRasterBand(1).WriteArray(data)
+        if nodata_value is not None:
+            temp_geotiff.GetRasterBand(1).SetNoDataValue(nodata_value)
+        temp_geotiff.SetGeoTransform(transform)
+        temp_geotiff.SetProjection(epsg_to_wkt(epsg_code))
+
+        driver = gdal.GetDriverByName('COG')
+        options = ['COMPRESS=LZW', 'OVERVIEW_RESAMPLING=AVERAGE', 'NUM_THREADS=ALL_CPUS', 'BIGTIFF=YES']
+        driver.CreateCopy(str(file_name), temp_geotiff, options=options)
+
+        del temp_geotiff  # How to close w/ gdal
+    return file_name
