+++ src/asf_tools/util.py
@@ -1,4 +1,6 @@
-from osgeo import gdal
+from typing import Tuple
+
+from osgeo import gdal, osr
 
 gdal.UseExceptions()
 
@@ -23,3 +25,45 @@ class GDALConfigManager:
     def __exit__(self, exc_type, exc_val, exc_tb):
         for key, value in self._previous_options.items():
             gdal.SetConfigOption(key, value)
+
+
+def get_epsg_code(info: dict) -> int:
+    """Get the EPSG code from a GDAL Info dictionary
+
+    Args:
+        info: The dictionary returned by a gdal.Info call
+
+    Returns:
+        epsg_code: The integer EPSG code
+    """
+    proj = osr.SpatialReference(info['coordinateSystem']['wkt'])
+    epsg_code = int(proj.GetAttrValue('AUTHORITY', 1))
+    return epsg_code
+
+
+def get_coordinates(info: dict) -> Tuple[int, int, int, int]:
+    """Get the corner coordinates from a GDAL Info dictionary
+
+    Args:
+        info: The dictionary returned by a gdal.Info call
+
+    Returns:
+        (west, south, east, north): the corner coordinates values
+    """
+    west, south = info['cornerCoordinates']['lowerLeft']
+    east, north = info['cornerCoordinates']['upperRight']
+    return west, south, east, north
+
+
+def epsg_to_wkt(epsg_code: int) -> str:
+    """Get the WKT representation of a projection from its EPSG code
+
+    Args:
+        epsg_code: The integer EPSG code
+
+    Returns:
+        wkt: The WKT representation of the projection
+    """
+    srs = osr.SpatialReference()
+    srs.ImportFromEPSG(epsg_code)
+    return srs.ExportToWkt()
