+++ tests/hydrosar/test_water_map.py
@@ -3,7 +3,7 @@ import pytest
 from osgeo_utils.gdalcompare import find_diff
 
 from asf_tools.hydrosar import water_map
-from asf_tools.composite import read_as_array
+from asf_tools.raster import read_as_array
 from asf_tools.tile import tile_array
