+++ tests/test_composite.py
@@ -2,41 +2,9 @@ import os
 
 import numpy as np
 import pytest
-from osgeo import gdal
 
+import asf_tools.raster
 from asf_tools import composite
-
-
-def test_get_epsg_code():
-    wkt = 'PROJCS["WGS 84 / UTM zone 54N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",141],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32654"]]'
-    info = {'coordinateSystem': {'wkt': wkt}}
-    assert composite.get_epsg_code(info) == 32654
-
-    wkt = 'PROJCS["WGS 84 / UTM zone 22N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-51],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32622"]]'
-    info = {'coordinateSystem': {'wkt': wkt}}
-    assert composite.get_epsg_code(info) == 32622
-
-    wkt = 'PROJCS["WGS 84 / UTM zone 33S",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",15],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",10000000],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32733"]]'
-    info = {'coordinateSystem': {'wkt': wkt}}
-    assert composite.get_epsg_code(info) == 32733
-
-    wkt = 'PROJCS["NAD83 / Alaska Albers",GEOGCS["NAD83",DATUM["North_American_Datum_1983",SPHEROID["GRS 1980",6378137,298.257222101,AUTHORITY["EPSG","7019"]],AUTHORITY["EPSG","6269"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4269"]],PROJECTION["Albers_Conic_Equal_Area"],PARAMETER["latitude_of_center",50],PARAMETER["longitude_of_center",-154],PARAMETER["standard_parallel_1",55],PARAMETER["standard_parallel_2",65],PARAMETER["false_easting",0],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","3338"]]'
-    info = {'coordinateSystem': {'wkt': wkt}}
-    assert composite.get_epsg_code(info) == 3338
-
-
-def test_epsg_to_wkt():
-    wkt = 'PROJCS["WGS 84 / UTM zone 54N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",141],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32654"]]'
-    assert composite.epsg_to_wkt(32654) == wkt
-
-    wkt = 'PROJCS["WGS 84 / UTM zone 22N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-51],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32622"]]'
-    assert composite.epsg_to_wkt(32622) == wkt
-
-    wkt = 'PROJCS["WGS 84 / UTM zone 33S",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",15],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",10000000],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32733"]]'
-    assert composite.epsg_to_wkt(32733) == wkt
-
-    wkt = 'PROJCS["NAD83 / Alaska Albers",GEOGCS["NAD83",DATUM["North_American_Datum_1983",SPHEROID["GRS 1980",6378137,298.257222101,AUTHORITY["EPSG","7019"]],AUTHORITY["EPSG","6269"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4269"]],PROJECTION["Albers_Conic_Equal_Area"],PARAMETER["latitude_of_center",50],PARAMETER["longitude_of_center",-154],PARAMETER["standard_parallel_1",55],PARAMETER["standard_parallel_2",65],PARAMETER["false_easting",0],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","3338"]]'
-    assert composite.epsg_to_wkt(3338) == wkt
 
 
 def test_get_target_epsg_code():
@@ -115,25 +83,6 @@ def test_get_full_extents():
     assert composite.get_full_extent(data) == (expected_upper_left, expected_lower_right, expected_geotransform)
 
 
-def test_write_cog(tmp_path):
-    outfile = tmp_path / 'out.tif'
-    data = np.ones((1024, 1024))
-    transform = [10.0, 0.0, 1.0, 20.0, 0.0, -1.0]
-    epsg_code = 4326
-
-    result = composite.write_cog(str(outfile), data, transform, epsg_code)
-    assert result == str(outfile)
-    assert outfile.exists()
-
-    info = gdal.Info(result, format='json')
-    assert info['geoTransform'] == transform
-    assert info['driverShortName'] == 'GTiff'
-    assert info['size'] == [1024, 1024]
-    assert 'overviews' in info['bands'][0]
-    assert info['metadata']['IMAGE_STRUCTURE']['LAYOUT'] == 'COG'
-    assert info['metadata']['IMAGE_STRUCTURE']['COMPRESSION'] == 'LZW'
-
-
 def test_make_composite(tmp_path):
     os.chdir(tmp_path)
     epsg_code = 32601
@@ -147,8 +96,8 @@ def test_make_composite(tmp_path):
         [1, 1, 1, 1],
         [1, 1, 1, 1],
     ])
-    composite.write_cog('first_data.tif', data, transform, epsg_code, nodata_value=0)
-    composite.write_cog('first_area.tif', area, transform, epsg_code)
+    asf_tools.raster.write_cog('first_data.tif', data, transform, epsg_code, nodata_value=0)
+    asf_tools.raster.write_cog('first_area.tif', area, transform, epsg_code)
 
     transform = [30.0, 30.0, 0.0, 30.0, 0.0, -30.0]
     data = np.array([
@@ -159,8 +108,8 @@ def test_make_composite(tmp_path):
         [1, 1, 3, 1],
         [1, 1, 2, 1],
     ])
-    composite.write_cog('second_data.tif', data, transform, epsg_code)
-    composite.write_cog('second_area.tif', area, transform, epsg_code)
+    asf_tools.raster.write_cog('second_data.tif', data, transform, epsg_code)
+    asf_tools.raster.write_cog('second_area.tif', area, transform, epsg_code)
 
     out_file, count_file = composite.make_composite('out', ['first_data.tif', 'second_data.tif'])
 
@@ -169,7 +118,7 @@ def test_make_composite(tmp_path):
     assert os.path.exists(out_file)
     assert os.path.exists(count_file)
 
-    data = np.nan_to_num(composite.read_as_array(out_file))
+    data = np.nan_to_num(asf_tools.raster.read_as_array(out_file))
     expected = np.array([
         [1, 1, 1,   1, 0],
         [1, 2, 1, 1.5, 3],
@@ -177,7 +126,7 @@ def test_make_composite(tmp_path):
     ])
     assert np.allclose(data, expected)
 
-    counts = composite.read_as_array(count_file)
+    counts = asf_tools.raster.read_as_array(count_file)
     expected = np.array([
         [1, 1, 1, 1, 0],
         [1, 2, 1, 2, 1],
