+++ tests/test_entrypoints.py
@@ -1,18 +1,18 @@
 def test_make_composite(script_runner):
-    ret = script_runner.run('make_composite', '-h')
+    ret = script_runner.run(['make_composite', '-h'])
     assert ret.success
 
 
 def test_water_map(script_runner):
-    ret = script_runner.run('water_map', '-h')
+    ret = script_runner.run(['water_map', '-h'])
     assert ret.success
 
 
 def test_make_hand(script_runner):
-    ret = script_runner.run('calculate_hand', '-h')
+    ret = script_runner.run(['calculate_hand', '-h'])
     assert ret.success
 
 
 def test_flood_map(script_runner):
-    ret = script_runner.run('flood_map', '-h')
+    ret = script_runner.run(['flood_map', '-h'])
     assert ret.success
