+++ tests/test_raster.py
@@ -1,6 +1,8 @@
 import numpy as np
 import pytest
 
+from osgeo import gdal
+
 from asf_tools import raster
 
 
@@ -51,3 +53,22 @@ def test_convert_scale_masked_arrays():
     a = raster.convert_scale(c, 'db', 'power')
     assert np.allclose(a.mask, [True, True, False, False, False])
     assert np.allclose(a, np.ma.MaskedArray([-1, 0, 1, 4, 9], mask=[True, True, False, False, False]))
+
+
+def test_write_cog(tmp_path):
+    outfile = tmp_path / 'out.tif'
+    data = np.ones((1024, 1024))
+    transform = [10.0, 0.0, 1.0, 20.0, 0.0, -1.0]
+    epsg_code = 4326
+
+    result = raster.write_cog(str(outfile), data, transform, epsg_code)
+    assert result == str(outfile)
+    assert outfile.exists()
+
+    info = gdal.Info(result, format='json')
+    assert info['geoTransform'] == transform
+    assert info['driverShortName'] == 'GTiff'
+    assert info['size'] == [1024, 1024]
+    assert 'overviews' in info['bands'][0]
+    assert info['metadata']['IMAGE_STRUCTURE']['LAYOUT'] == 'COG'
+    assert info['metadata']['IMAGE_STRUCTURE']['COMPRESSION'] == 'LZW'
